+++ ChiffreEnLettre/Utilitaire.h
@@ -2,6 +2,6 @@
 
 int nombreDeChiffres(int);
 
-int* separer(int, int, int*);
+int* separer(long long int, int, int*);
 
 char* joindre(char*, char*, char*);
